% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aa_environments.r
\name{nb}
\alias{nb}
\title{Norwegian characters in unicode}
\usage{
nb
}
\description{
A named list of Norwegian special characters as unicode strings,
for use where literal non-ASCII characters are inconvenient.
Elements: \code{AA}/\code{aa} (Aa/aa), \code{OE}/\code{oe} (Oe/oe), \code{AE}/\code{ae} (Ae/ae).
}
\examples{
print(csdata::nb)
csdata::nb$AA  # uppercase Aa
}
